+++ lib/generators/react_on_rails/base_generator.rb
@@ -118,9 +118,13 @@ module ReactOnRails
         return unless options.server_rendering?
         base_path = "base/server_rendering/"
         %w(client/webpack.server.rails.config.js
-           client/app/bundles/HelloWorld/startup/serverRegistration.jsx).each do |file|
+           client/app/bundles/HelloWorld/startup/serverRegistration.jsx
+           client/node/package.json
+           client/node/server.js).each do |file|
           copy_file(base_path + file, file)
         end
+
+        copy_file("base/base/lib/tasks/load_test.rake", "lib/tasks/load_test.rake")
       end
 
       def template_assets_rake_file
