+++ lib/generators/react_on_rails/templates/base/base/config/initializers/react_on_rails.rb.tt
@@ -29,7 +29,7 @@ ReactOnRails.configure do |config|
   # For server rendering. This can be set to false so that server side messages are discarded.
   # Default is true. Be cautious about turning this off.
   config.replay_console = true
-  # Default is true. Logs server rendering messags to Rails.logger.info
+  # Default is true. Logs server rendering messages to Rails.logger.info
   config.logging_on_server = true
 
   # The following options can be overriden by passing to the helper method:
@@ -40,4 +40,7 @@ ReactOnRails.configure do |config|
   config.trace = Rails.env.development?
   # Default is false, enable if your content security policy doesn't include `style-src: 'unsafe-inline'`
   config.skip_display_none = false
+
+  # The server render method - either ExecJS or NodeJS
+  config.server_render_method = "ExecJS"
 end
