+++ lib/generators/react_on_rails/templates/base/base/lib/tasks/load_test.rake
@@ -1,0 +1,8 @@
+namespace :load_test do
+  desc "Load test with apache benchmark"
+  task :run, [:url, :count] do |_, args|
+    url = args[:url] || "http://localhost:3000/hello_world"
+    count = args[:count] || 500
+    system("ab -c 10 -n #{count} #{url}")
+  end
+end
