+++ lib/generators/react_on_rails/templates/base/server_rendering/client/node/package.json
@@ -1,0 +1,10 @@
+{
+    "name": "react_on_rails_node",
+    "version": "0.0.0",
+    "private": true,
+    "scripts": {
+        "start": "node ./server.js -s server-bundle.js"
+    },
+    "dependencies": {
+    }
+}
