+++ lib/generators/react_on_rails/templates/dev_tests/spec/rails_helper.rb
@@ -36,6 +36,7 @@ Capybara.javascript_driver = :poltergeist
 RSpec.configure do |config|
   # Ensure that if we are running js tests, we are using latest webpack assets
   # This will use the defaults of :js and :server_rendering meta tags
+  ReactOnRails::TestHelper.launch_node if ReactOnRails.configuration.server_render_method == "NodeJS"
   ReactOnRails::TestHelper.configure_rspec_to_compile_assets(config)
 
   # Remove this line if you"re not using ActiveRecord or ActiveRecord fixtures
