+++ lib/react_on_rails.rb
@@ -16,3 +16,4 @@ require "react_on_rails/test_helper/webpack_assets_compiler"
 require "react_on_rails/test_helper/webpack_assets_status_checker"
 require "react_on_rails/test_helper/webpack_process_checker"
 require "react_on_rails/test_helper/ensure_assets_compiled"
+require "react_on_rails/test_helper/node_process_launcher"
