+++ lib/react_on_rails/configuration.rb
@@ -56,7 +56,8 @@ module ReactOnRails
       server_renderer_timeout: 20,
       skip_display_none: false,
       webpack_generated_files: [],
-      rendering_extension: nil
+      rendering_extension: nil,
+      server_render_method: ""
     )
   end
 
@@ -66,7 +67,7 @@ module ReactOnRails
                   :logging_on_server, :server_renderer_pool_size,
                   :server_renderer_timeout, :raise_on_prerender_error,
                   :skip_display_none, :generated_assets_dirs, :generated_assets_dir,
-                  :webpack_generated_files, :rendering_extension
+                  :webpack_generated_files, :rendering_extension, :server_render_method
 
     def initialize(server_bundle_js_file: nil, prerender: nil, replay_console: nil,
                    trace: nil, development_mode: nil,
@@ -74,7 +75,7 @@ module ReactOnRails
                    server_renderer_timeout: nil, raise_on_prerender_error: nil,
                    skip_display_none: nil, generated_assets_dirs: nil,
                    generated_assets_dir: nil, webpack_generated_files: nil,
-                   rendering_extension: nil)
+                   rendering_extension: nil, server_render_method: nil)
       self.server_bundle_js_file = server_bundle_js_file
       self.generated_assets_dirs = generated_assets_dirs
       self.generated_assets_dir = generated_assets_dir
@@ -97,6 +98,8 @@ module ReactOnRails
 
       self.webpack_generated_files = webpack_generated_files
       self.rendering_extension = rendering_extension
+
+      self.server_render_method = server_render_method
     end
   end
 end
